rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.2.2",
    "gitpython>=3.1.44",
    "pandas>=2.2.3",
    "pyarrow>=19.0.1",
    "sentence-transformers>=4.0.2",
]

[tool.mypy]
# Ignore missing stubs for specific third-party libraries
[[tool.mypy.overrides]]
module = [
    "sentence_transformers.*",
    "git.*", #念のためgitも追加
    "duckdb.*" #念のためduckdbも追加
]
ignore_missing_imports = true
